\name{suma}
\alias{suma}
\title{suma de dos variables}

\usage{
suma(x,y)
}

\arguments{
  \item{a}{
  Es la primera variable de la suma
  }
  \item{b}{
  Es la segunda variable de la suma
  }
}

\value{
Retorna un valor numerico, que es el resultado de la funcion suma \code{x}.
}

\description{
La funcion suma es una operacion aritmetica basica en combinar o añadir dos números para obtener una cantidad total.
}

\examples{
# Suma de 10 entre 5
suma(10, 5)
}

\references{
{Hansen, N., Müller, S. D., & Koumoutsakos, P. (2003). \emph{Reducing the time complexity of the derandomized evolution strategy with covariance matrix adaptation (CMA-ES)}. Evolutionary Computation, 11(1), 1-18.}
}
